+++ x/stakeibc/keeper/deposit_records_test.go
@@ -207,89 +207,6 @@ func (s *KeeperTestSuite) SetupDepositRecords() DepositRecordsTestCase {
 	}
 }
 
-func (s *KeeperTestSuite) TestCreateDepositRecordsForEpoch_Successful() {
-	// Set host zones
-	hostZones := []stakeibctypes.HostZone{
-		{
-			ChainId:   "HOST1",
-			HostDenom: "denom1",
-		},
-		{
-			ChainId:   "HOST2",
-			HostDenom: "denom2",
-		},
-		{
-			ChainId:   "HOST3",
-			HostDenom: "denom3",
-		},
-	}
-	for _, hostZone := range hostZones {
-		s.App.StakeibcKeeper.SetHostZone(s.Ctx, hostZone)
-	}
-
-	// Create depoist records for two epochs
-	s.App.StakeibcKeeper.CreateDepositRecordsForEpoch(s.Ctx, 1)
-	s.App.StakeibcKeeper.CreateDepositRecordsForEpoch(s.Ctx, 2)
-
-	expectedDepositRecords := []recordstypes.DepositRecord{
-		// Epoch 1
-		{
-			Id:                 0,
-			Amount:             sdkmath.ZeroInt(),
-			Denom:              "denom1",
-			HostZoneId:         "HOST1",
-			Status:             recordstypes.DepositRecord_TRANSFER_QUEUE,
-			DepositEpochNumber: 1,
-		},
-		{
-			Id:                 1,
-			Amount:             sdkmath.ZeroInt(),
-			Denom:              "denom2",
-			HostZoneId:         "HOST2",
-			Status:             recordstypes.DepositRecord_TRANSFER_QUEUE,
-			DepositEpochNumber: 1,
-		},
-		{
-			Id:                 2,
-			Amount:             sdkmath.ZeroInt(),
-			Denom:              "denom3",
-			HostZoneId:         "HOST3",
-			Status:             recordstypes.DepositRecord_TRANSFER_QUEUE,
-			DepositEpochNumber: 1,
-		},
-		// Epoch 2
-		{
-			Id:                 3,
-			Amount:             sdkmath.ZeroInt(),
-			Denom:              "denom1",
-			HostZoneId:         "HOST1",
-			Status:             recordstypes.DepositRecord_TRANSFER_QUEUE,
-			DepositEpochNumber: 2,
-		},
-		{
-			Id:                 4,
-			Amount:             sdkmath.ZeroInt(),
-			Denom:              "denom2",
-			HostZoneId:         "HOST2",
-			Status:             recordstypes.DepositRecord_TRANSFER_QUEUE,
-			DepositEpochNumber: 2,
-		},
-		{
-			Id:                 5,
-			Amount:             sdkmath.ZeroInt(),
-			Denom:              "denom3",
-			HostZoneId:         "HOST3",
-			Status:             recordstypes.DepositRecord_TRANSFER_QUEUE,
-			DepositEpochNumber: 2,
-		},
-	}
-
-	// Confirm deposit records
-	actualDepositRecords := s.App.RecordsKeeper.GetAllDepositRecord(s.Ctx)
-	s.Require().Equal(len(expectedDepositRecords), len(actualDepositRecords), "number of deposit records")
-	s.Require().Equal(expectedDepositRecords, actualDepositRecords, "deposit records")
-}
-
 // Helper function to check the state after transferring deposit records
 // This assumes the last X transfers failed
 func (s *KeeperTestSuite) CheckStateAfterTransferringDepositRecords(tc DepositRecordsTestCase, numTransfersFailed int) {
